+++ src/main/java/dev/astranfalio/teioc/service/TopicDataService.java
@@ -6,7 +6,7 @@ import jakarta.validation.Validator;
 import org.springframework.stereotype.Service;
 
 @Service
-public class TopicDataService extends AbstractDataService<TopicEntity, Long, TopicRepository> {
+public class TopicDataService extends AbstractDataService<TopicEntity, Integer, TopicRepository> {
     public TopicDataService(TopicRepository repository, Validator validator) {
         super(repository, validator);
     }
